+++ frontend/portals/CartCheckoutButton/index.jsx
@@ -6,22 +6,33 @@ import RechargeCheckoutButton from '../../components/RechargeCheckoutButton';
 /**
  * @returns {JSX}
  */
-const CartCheckoutButton = ({ cartToken, disabled, children }) => {
-  if (!cartToken) {
+const CartCheckoutButton = ({
+  cartToken,
+  disabled,
+  children,
+  isTokenFetching,
+  isError,
+}) => {
+  if (!cartToken && !isTokenFetching && !isError) {
     return children;
   }
-  return (<RechargeCheckoutButton cartToken={cartToken} disabled={disabled} />);
+
+  return (<RechargeCheckoutButton cartToken={cartToken} disabled={disabled || isError} />);
 };
 
 CartCheckoutButton.propTypes = {
   children: PropTypes.node.isRequired,
   cartToken: PropTypes.string,
   disabled: PropTypes.bool,
+  isError: PropTypes.bool,
+  isTokenFetching: PropTypes.bool,
 };
 
 CartCheckoutButton.defaultProps = {
   cartToken: null,
   disabled: false,
+  isError: false,
+  isTokenFetching: false,
 };
 
 export default connect(CartCheckoutButton);